Move match functionality into match*.go
